1.0.62

+++ src/main/java/com/jbnucpu/www/service/DeleteService.java
@@ -25,12 +25,12 @@ public class DeleteService {
             System.out.println("삭제 실패: 로그인 한 사용자만 삭제 가능");
             return false;
         }
-        // 작성자만 삭제 가능
-//        NoticeEntity noticeEntity = noticeRepository.findById(id).orElseThrow(()->new NoSuchElementException("해당 id에 대한 게시글을 찾을 수 없습니다"));
-//        if(!authService.getUsername().equals(noticeEntity.getStudentNumber())){
-//            System.out.println("삭제 실패: 작성자만 삭제 가능");
-//            return false;
-//        }
+
+        NoticeEntity noticeEntity = noticeRepository.findById(id).orElseThrow(()->new NoSuchElementException("해당 id에 대한 게시글을 찾을 수 없습니다"));
+        if(!authService.getUsername().equals(noticeEntity.getUserEntity().getStudentnumber())){
+            System.out.println("삭제 실패: 작성자만 삭제 가능");
+            return false;
+        }
 
         noticeRepository.deleteById(id);
